end of the day slaying i'm done for now

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: mapale <mapale@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/08 09:06:30 by mapale            #+#    #+#             */
-/*   Updated: 2023/11/08 10:04:43 by mapale           ###   ########.fr       */
+/*   Updated: 2023/11/09 17:37:04 by mapale           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void    *ft_memcpy(void *dest, const void *src, size_t n)
 {
     char *tmp_dest;
     char *tmp_src;
-    int     i;
+    size_t     i;
 
     tmp_dest = (char *)dest;
     tmp_src = (char *)src;
