+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: mapale <mapale@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/08 13:16:19 by mapale            #+#    #+#             */
-/*   Updated: 2023/11/08 13:20:22 by mapale           ###   ########.fr       */
+/*   Updated: 2023/11/09 17:48:49 by mapale           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-    int i;
+    size_t	i;
 
     i = 0;
     while (s1[i] && s2[i] && i <= n)
