+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: mapale <mapale@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/08 16:29:48 by mapale            #+#    #+#             */
-/*   Updated: 2023/11/08 16:37:07 by mapale           ###   ########.fr       */
+/*   Updated: 2023/11/09 17:51:07 by mapale           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	int		i;
+	size_t		i;
 	char	*tab;
 
 	i = 0;
